reverted changes to logger.cpp:194 and logger.h:121
This removes changes to the stack log that indents the output and kept track of what the last stack push was. The problem was that the logger is called from many different threads at the same time. Static variables were used to keep track of this, different threads would change the static variable.

Instead of static variables thread specific data needs to be used.